vertexcover/requirements.txt
Django==3.0.6
django-allauth==0.41.0
django-extensions==2.2.9
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
httplib2==0.9.2
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
social-auth-app-django==3.1.0
social-auth-core==3.3.3
urllib3==1.25.9
ortools==7.6.7691
